# server/internal/api/openapi/model_flashcard.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Flashcard API
 *
 * API for managing flashcards in a learning application
 *
 * API version: 1.0.0
 * Contact: support@example.com
 */

package api


import (
	"time"
)



type Flashcard struct {

	Id int64 `json:"id"`

	Front string `json:"front"`

	Back string `json:"back"`

	Category string `json:"category,omitempty"`

	CreatedAt time.Time `json:"created_at"`

	UpdatedAt time.Time `json:"updated_at"`
}

// AssertFlashcardRequired checks if the required fields are not zero-ed
func AssertFlashcardRequired(obj Flashcard) error {
	elements := map[string]interface{}{
		"id": obj.Id,
		"front": obj.Front,
		"back": obj.Back,
		"created_at": obj.CreatedAt,
		"updated_at": obj.UpdatedAt,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertFlashcardConstraints checks if the values respects the defined constraints
func AssertFlashcardConstraints(obj Flashcard) error {
	return nil
}