# server/internal/api/openapi/model_flashcard_create_request.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Flashcard API
 *
 * API for managing flashcards in a learning application
 *
 * API version: 1.0.0
 * Contact: support@example.com
 */

package api




type FlashcardCreateRequest struct {

	Front string `json:"front"`

	Back string `json:"back"`

	Category string `json:"category,omitempty"`
}

// AssertFlashcardCreateRequestRequired checks if the required fields are not zero-ed
func AssertFlashcardCreateRequestRequired(obj FlashcardCreateRequest) error {
	elements := map[string]interface{}{
		"front": obj.Front,
		"back": obj.Back,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertFlashcardCreateRequestConstraints checks if the values respects the defined constraints
func AssertFlashcardCreateRequestConstraints(obj FlashcardCreateRequest) error {
	return nil
}