# server/internal/api/openapi/model_flashcard_update_request.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Flashcard API
 *
 * API for managing flashcards in a learning application
 *
 * API version: 1.0.0
 * Contact: support@example.com
 */

package api




type FlashcardUpdateRequest struct {

	Front string `json:"front,omitempty"`

	Back string `json:"back,omitempty"`

	Category string `json:"category,omitempty"`
}

// AssertFlashcardUpdateRequestRequired checks if the required fields are not zero-ed
func AssertFlashcardUpdateRequestRequired(obj FlashcardUpdateRequest) error {
	return nil
}

// AssertFlashcardUpdateRequestConstraints checks if the values respects the defined constraints
func AssertFlashcardUpdateRequestConstraints(obj FlashcardUpdateRequest) error {
	return nil
}